crear endpoint /incidencias - depuracion de codigo

+++ backend/src/main/java/com/callcenter/NoCountry/DTO/IncidenciaDTO.java
@@ -6,8 +6,8 @@ import java.time.LocalDateTime;
 
 @Data
 public class IncidenciaDTO {
-    private Integer idCliente;
-    private Integer idServicio;
+    private Long idCliente;
+    private Long idServicio;
     private int prioridad;
     private String estado;
     private String descripcion;
